+++ lib/authmac/hmac_checker.rb
@@ -2,10 +2,16 @@ require 'openssl'
 
 module Authmac
   class HmacChecker
-    def initialize(secret, parameter_separator = '|', digest_function = 'sha1')
+    # @param message_format [symbol] `:values` or `:json`.
+    #   `:json` will use a sorted json string to sign.
+    #   `:values` will use the sorted values separated by `parameter_separator` to sign.
+    def initialize(secret, message_format: :values,
+                           parameter_separator: '|',
+                           digest_function: 'sha1')
       @secret = secret
       @digest = digest_function
       @separator = parameter_separator
+      @message_format = message_format
       fail Authmac::SecretError, 'secret too short, see rfc2104' unless @secret.bytes.size >= digester.digest_length * 2
     end
 
@@ -28,12 +34,35 @@ module Authmac
     end
 
     def message_string(hash)
-      hash_values_sorted_by_key(hash).join(@separator)
+      fail ArgumentError, 'hash arg not a hash' unless hash.is_a? Hash
+
+      case @message_format
+      when :values
+        hash_values_sorted_by_key(hash).flatten.join(@separator)
+      when :json
+        require 'json'
+        JSON.generate(params_sorted_by_key(hash))
+      else
+        fail ArgumentError, 'unknown message_format'
+      end
     end
 
     def hash_values_sorted_by_key(hash)
       hash.sort_by {|key, value| key }.map(&:last)
     end
+
+    # stringifies and sorts hashes by key at all levels.
+    def params_sorted_by_key(params)
+      case params
+      when Hash
+        params.map     { |k, v| [k.to_s, params_sorted_by_key(v)] }
+              .sort_by { |k, v| k }
+              .to_h
+      when Array
+        params.map { |val| params_sorted_by_key(val) }
+      else
+        params
+      end
+    end
   end
 end
-