+++ spec/authmac/hmac_checker_spec.rb
@@ -2,11 +2,11 @@ require 'authmac/hmac_checker'
 
 module Authmac
   describe HmacChecker do
-    let(:checker) { HmacChecker.new('very secret random key of sufficient size', '|', 'sha1') }
+    let(:checker) { HmacChecker.new('very secret random key of sufficient size') }
 
     it 'raises an error for a secret shorter than the hmac output' do
       expect {
-        HmacChecker.new('way too short key', '|', 'sha1')
+        HmacChecker.new('way too short key')
       }.to raise_error SecretError, 'secret too short, see rfc2104'
     end
 
@@ -43,6 +43,42 @@ module Authmac
         it 'sorts hash values based on their keys' do
           expect(checker.validate({second: 'another', first: 'parameter'},
                                   hmacify('parameter|another'))).to be_truthy
+        end
+      end
+
+      context 'with json message_string' do
+        let(:checker) { HmacChecker.new('very secret random key of sufficient size', message_format: :json) }
+
+        it 'creates a sorted json string' do
+          hash = {first: '1', third: '6', second: {b: ['3', {c: '4', d: '5'}], a: '2'}}
+          expect(checker.send :message_string, hash)
+            .to eq '{"first":"1","second":{"a":"2","b":["3",{"c":"4","d":"5"}]},"third":"6"}'
+        end
+
+        context 'for a hash with nested hashes' do
+          it 'succeeds with correct hmac' do
+            expect(checker.validate({first: '1', second: {a: '2', b: '3'}, third: '4'},
+                                    hmacify('{"first":"1","second":{"a":"2","b":"3"},"third":"4"}'))).to be_truthy
+          end
+
+          it 'sorts hash values based on their keys' do
+            expect(checker.validate({first: '1', third: '4', second: {b: '3', a: '2'}},
+                                    hmacify('{"first":"1","second":{"a":"2","b":"3"},"third":"4"}'))).to be_truthy
+          end
+        end
+
+        context 'for a hash with nested hashes and arrays' do
+          it 'succeeds with correct hmac' do
+            expect(checker.validate({first: '1', second: {a: '2', b: ['3', '4']}, third: '5'},
+                                    hmacify('{"first":"1","second":{"a":"2","b":["3","4"]},"third":"5"}')))
+                          .to be_truthy
+          end
+
+          it 'sorts hash values based on their keys' do
+            expect(checker.validate({first: '1', third: '6', second: {b: ['3', {c: '4', d: '5'}], a: '2'}},
+                                    hmacify( '{"first":"1","second":{"a":"2","b":["3",{"c":"4","d":"5"}]},"third":"6"}')))
+                          .to be_truthy
+          end
         end
       end
     end