+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unpublished Version 2.0.0
+
+* Allow nested hashes/arrays to be signed.
+* Add :json message_format to sign a sorted json string, instead of just values.
+
 ## Version 1.0.4 / 2014-06-20
 
 * Added HmacChecker#with_signature which returns the given hash with the HMAC merged in.